Refactored code in different files
