+++ main.go
@@ -1,49 +1,11 @@
 package main
 
 import (
-	"fmt"
 	"log"
 	"net/http"
-
-	"github.com/gorilla/mux"
-	"time"
-	"encoding/json"
 )
 
-type Todo struct {
-	Name      string    `json:"name"`
-	Completed bool      `json:"completed"`
-	Due       time.Time `json:"due"`
-}
-
-type Todos []Todo
-
 func main() {
-	router := mux.NewRouter().StrictSlash(true)
-
-	router.HandleFunc("/", Index)
-	router.HandleFunc("/todos", TodoIndex)
-	router.HandleFunc("/todos/{todoId}", TodoShow)
-
+	router := NewRouter()
 	log.Fatal(http.ListenAndServe(":8080", router))
-}
-
-func Index(w http.ResponseWriter, r *http.Request) {
-	fmt.Fprintln(w, "Welcome amigo!")
-}
-
-func TodoIndex(w http.ResponseWriter, r *http.Request) {
-
-	todos := Todos{
-		Todo{Name: "Write Presentation"},
-		Todo{Name: "Host meetup"},
-	}
-
-	json.NewEncoder(w).Encode(todos)
-}
-
-func TodoShow(w http.ResponseWriter, r *http.Request) {
-	vars := mux.Vars(r)
-	todoId := vars["todoId"]
-	fmt.Fprintln(w, "Todo Show: ", todoId)
 }
