+++ todos.go
@@ -1,0 +1,11 @@
+package main
+
+import "time"
+
+type Todo struct {
+	Name      string    `json:"name"`
+	Completed bool      `json:"completed"`
+	Due       time.Time `json:"due"`
+}
+
+type Todos []Todo
